[refactor] 1:1채팅 - 채팅방id를 useRef에 지정, props로 넘겨주기
